+++ primefaces/src/main/java/org/primefaces/util/CurrencyValidator.java
@@ -23,10 +23,7 @@
  */
 package org.primefaces.util;
 
-import java.io.Serializable;
-import java.math.BigDecimal;
-import java.math.RoundingMode;
-import java.text.*;
+import java.text.DecimalFormat;
 
 /**
  * <p>
@@ -42,15 +39,15 @@ import java.text.*;
  * Modified from Commons Validator to fit PF needs
  * @see <a href="https://github.com/apache/commons-validator">Commons Validator</a>
  */
-public class CurrencyValidator implements Serializable {
+public class CurrencyValidator extends BigDecimalValidator {
+    /** DecimalFormat's currency symbol */
+    public static final char CURRENCY_SYMBOL = '\u00A4';
+    public static final String CURRENCY_SYMBOL_STR = Character.toString(CURRENCY_SYMBOL);
 
     private static final long serialVersionUID = -4201640771171486514L;
 
     private static final CurrencyValidator VALIDATOR = new CurrencyValidator();
 
-    /** DecimalFormat's currency symbol */
-    private static final char CURRENCY_SYMBOL = '\u00A4';
-    private static final String CURRENCY_SYMBOL_STR = Character.toString(CURRENCY_SYMBOL);
     /** Space hack to fix Brazilian Real and maybe others */
     private static final char NON_BREAKING_SPACE = '\u00A0';
     private static final String NON_BREAKING_SPACE_STR = Character.toString(NON_BREAKING_SPACE);
@@ -66,18 +63,6 @@ public class CurrencyValidator implements Serializable {
 
     /**
      * <p>
-     * Validate/convert a <code>BigDecimal</code> using the specified <code>Locale</code>.
-     *
-     * @param value The value validation is being performed on.
-     * @param format The format to use for the number format, system default if null.
-     * @return The parsed <code>BigDecimal</code> if valid or <code>null</code> if invalid.
-     */
-    public BigDecimal validate(String value, DecimalFormat format) {
-        return (BigDecimal) parse(value, format);
-    }
-
-    /**
-     * <p>
      * Parse the value with the specified <code>Format</code>.
      * </p>
      * <p>
@@ -89,7 +74,8 @@ public class CurrencyValidator implements Serializable {
      * @param formatter The Format to parse the value with.
      * @return The parsed value if valid or <code>null</code> if invalid.
      */
-    protected Object parse(String value, DecimalFormat formatter) {
+    @Override
+    protected Number parse(String value, DecimalFormat formatter) {
         // check and replace currency symbol
         if (value.indexOf(CURRENCY_SYMBOL) >= 0) {
             value = value.replace(CURRENCY_SYMBOL_STR, formatter.getDecimalFormatSymbols().getCurrencySymbol());
@@ -101,7 +87,7 @@ public class CurrencyValidator implements Serializable {
         }
 
         // Initial parse of the value
-        Number parsedValue = parseFormat(value, formatter);
+        Number parsedValue = super.parse(value, formatter);
         if (parsedValue != null) {
             return parsedValue;
         }
@@ -117,106 +103,8 @@ public class CurrencyValidator implements Serializable {
             }
             DecimalFormat copyFormatter = (DecimalFormat) formatter.clone();
             copyFormatter.applyPattern(buffer.toString());
-            parsedValue = parseFormat(value, copyFormatter);
+            parsedValue = super.parse(value, copyFormatter);
         }
         return parsedValue;
     }
-
-    /**
-     * <p>
-     * Parse the value with the specified <code>Format</code>.
-     * </p>
-     *
-     * @param value The value to be parsed.
-     * @param formatter The Format to parse the value with.
-     * @return The parsed value if valid or <code>null</code> if invalid.
-     */
-    protected Number parseFormat(String value, DecimalFormat formatter) {
-        ParsePosition pos = new ParsePosition(0);
-        Number parsedValue = formatter.parse(value, pos);
-        if (pos.getErrorIndex() > -1) {
-            return null;
-        }
-
-        if (pos.getIndex() < value.length()) {
-            return null;
-        }
-
-        if (parsedValue != null) {
-            if (Double.isInfinite(parsedValue.doubleValue())) {
-                return null;
-            }
-            parsedValue = processParsedValue(parsedValue, formatter);
-        }
-
-        return parsedValue;
-
-    }
-
-    /**
-     * Convert the parsed value to a <code>BigDecimal</code>.
-     *
-     * @param value The parsed <code>Number</code> object created.
-     * @param formatter The Format used to parse the value with.
-     * @return The parsed <code>Number</code> converted to a <code>BigDecimal</code>.
-     */
-    protected BigDecimal processParsedValue(Number value, DecimalFormat formatter) {
-        BigDecimal decimal;
-        if (value instanceof Long) {
-            decimal = BigDecimal.valueOf((Long) value);
-        }
-        else {
-            decimal = new BigDecimal(value.toString());
-        }
-
-        int scale = determineScale(formatter);
-        if (scale >= 0) {
-            decimal = decimal.setScale(scale, RoundingMode.DOWN);
-        }
-
-        return decimal;
-    }
-
-    /**
-     * <p>
-     * Returns the <i>multiplier</i> of the <code>NumberFormat</code>.
-     * </p>
-     *
-     * @param format The <code>NumberFormat</code> to determine the multiplier of.
-     * @return The multiplying factor for the format.
-     */
-    protected int determineScale(NumberFormat format) {
-        int minimumFraction = format.getMinimumFractionDigits();
-        int maximumFraction = format.getMaximumFractionDigits();
-        if (minimumFraction != maximumFraction) {
-            return -1;
-        }
-        int scale = minimumFraction;
-        if (format instanceof DecimalFormat) {
-            int multiplier = ((DecimalFormat) format).getMultiplier();
-            if (multiplier == 100) { // CHECKSTYLE IGNORE MagicNumber
-                scale += 2; // CHECKSTYLE IGNORE MagicNumber
-            }
-            else if (multiplier == 1000) { // CHECKSTYLE IGNORE MagicNumber
-                scale += 3; // CHECKSTYLE IGNORE MagicNumber
-            }
-        }
-        return scale;
-    }
-
-    /**
-     * <p>
-     * Returns a <code>String</code> representing the Excel pattern for this currency.
-     * </p>
-     *
-     * @param format The format a <code>NumberFormat</code> is required for.
-     * @return The <code>String</code> pattern for using in Excel format.
-     */
-    public String getExcelPattern(DecimalFormat format) {
-        String pattern = format.toLocalizedPattern();
-        pattern =  pattern.replace(CURRENCY_SYMBOL_STR, "\"" + format.getDecimalFormatSymbols().getCurrencySymbol() + "\"");
-        String[] patterns = pattern.split(";");
-        return patterns[0];
-    }
-
 }